Checkstyle: Fix top-level class violations
