+++ src/main/java/games/strategy/engine/history/EventChild.java
@@ -26,20 +26,3 @@ public class EventChild extends HistoryNode implements Renderable {
     return new EventChildWriter(m_text, m_renderingData);
   }
 }
-
-
-class EventChildWriter implements SerializationWriter {
-  private static final long serialVersionUID = -7143658060171295697L;
-  private final String m_text;
-  private final Object m_renderingData;
-
-  public EventChildWriter(final String text, final Object renderingData) {
-    m_text = text;
-    m_renderingData = renderingData;
-  }
-
-  @Override
-  public void write(final HistoryWriter writer) {
-    writer.addChildToEvent(new EventChild(m_text, m_renderingData));
-  }
-}
