+++ src/main/java/games/strategy/engine/history/History.java
@@ -1,6 +1,5 @@
 package games.strategy.engine.history;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
@@ -179,78 +178,5 @@ public class History extends DefaultTreeModel {
 
   GameData getGameData() {
     return m_data;
-  }
-}
-
-
-/**
- * DefaultTreeModel is not serializable across jdk versions
- * Instead we use an instance of this class to store our data.
- */
-class SerializedHistory implements Serializable {
-  private static final long serialVersionUID = -5808427923253751651L;
-  private final List<SerializationWriter> m_Writers = new ArrayList<>();
-  private final GameData m_data;
-
-  public SerializedHistory(final History history, final GameData data, final List<Change> changes) {
-    m_data = data;
-    int changeIndex = 0;
-    final Enumeration<?> enumeration = ((DefaultMutableTreeNode) history.getRoot()).preorderEnumeration();
-    enumeration.nextElement();
-    while (enumeration.hasMoreElements()) {
-      final HistoryNode node = (HistoryNode) enumeration.nextElement();
-      // write the changes to the start of the node
-      if (node instanceof IndexedHistoryNode) {
-        while (changeIndex < ((IndexedHistoryNode) node).getChangeStartIndex()) {
-          m_Writers.add(new ChangeSerializationWriter(changes.get(changeIndex)));
-          changeIndex++;
-        }
-      }
-      // write the node itself
-      m_Writers.add(node.getWriter());
-    }
-    // write out remaining changes
-    while (changeIndex < changes.size()) {
-      m_Writers.add(new ChangeSerializationWriter(changes.get(changeIndex)));
-      changeIndex++;
-    }
-  }
-
-  public Object readResolve() {
-    final History history = new History(m_data);
-    final HistoryWriter historyWriter = history.getHistoryWriter();
-    for (final SerializationWriter element : m_Writers) {
-      element.write(historyWriter);
-    }
-    return history;
-  }
-}
-
-
-class RootHistoryNode extends HistoryNode {
-  private static final long serialVersionUID = 625147613043836829L;
-
-  public RootHistoryNode(final String title) {
-    super(title);
-  }
-
-  @Override
-  public SerializationWriter getWriter() {
-    throw new IllegalStateException("Not implemented");
-  }
-}
-
-
-class ChangeSerializationWriter implements SerializationWriter {
-  private static final long serialVersionUID = -3802807345707883606L;
-  private final Change aChange;
-
-  public ChangeSerializationWriter(final Change change) {
-    aChange = change;
-  }
-
-  @Override
-  public void write(final HistoryWriter writer) {
-    writer.addChange(aChange);
   }
 }
