+++ src/main/java/games/strategy/engine/history/Round.java
@@ -18,18 +18,3 @@ public class Round extends IndexedHistoryNode {
     return new RoundHistorySerializer(m_RoundNo);
   }
 }
-
-
-class RoundHistorySerializer implements SerializationWriter {
-  private static final long serialVersionUID = 9006488114384654514L;
-  private final int m_roundNo;
-
-  public RoundHistorySerializer(final int roundNo) {
-    m_roundNo = roundNo;
-  }
-
-  @Override
-  public void write(final HistoryWriter writer) {
-    writer.startNextRound(m_roundNo);
-  }
-}
