+++ src/main/java/games/strategy/triplea/delegate/AbstractEndTurnDelegate.java
@@ -582,12 +582,3 @@ public abstract class AbstractEndTurnDelegate extends BaseTripleADelegate implem
     };
   }
 }
-
-
-class EndTurnExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = -3939461840835898284L;
-  Serializable superState;
-  // add other variables here:
-  public boolean m_needToInitialize;
-  public boolean m_hasPostedTurnSummary;
-}
