+++ src/main/java/games/strategy/triplea/delegate/BaseDelegateState.java
@@ -1,0 +1,9 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class BaseDelegateState implements Serializable {
+  private static final long serialVersionUID = 7130686697155151908L;
+  public boolean m_startBaseStepsFinished = false;
+  public boolean m_endBaseStepsFinished = false;
+}
