+++ src/main/java/games/strategy/triplea/delegate/BaseTripleADelegate.java
@@ -128,10 +128,3 @@ public abstract class BaseTripleADelegate extends AbstractDelegate {
     return (ITripleAPlayer) bridge.getRemotePlayer(player);
   }
 }
-
-
-class BaseDelegateState implements Serializable {
-  private static final long serialVersionUID = 7130686697155151908L;
-  public boolean m_startBaseStepsFinished = false;
-  public boolean m_endBaseStepsFinished = false;
-}
