+++ src/main/java/games/strategy/triplea/delegate/BattleDelegate.java
@@ -1572,21 +1572,3 @@ public class BattleDelegate extends BaseTripleADelegate implements IBattleDelega
     return m_currentBattle;
   }
 }
-
-
-class BattleExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = 7899007486408723505L;
-  Serializable superState;
-  // add other variables here:
-  BattleTracker m_battleTracker = new BattleTracker();
-  // public OriginalOwnerTracker m_originalOwnerTracker = new OriginalOwnerTracker();
-  public boolean m_needToInitialize;
-  boolean m_needToScramble;
-  boolean m_needToKamikazeSuicideAttacks;
-  boolean m_needToClearEmptyAirBattleAttacks;
-  boolean m_needToAddBombardmentSources;
-  boolean m_needToRecordBattleStatistics;
-  boolean m_needToCheckDefendingPlanesCanLand;
-  boolean m_needToCleanup;
-  IBattle m_currentBattle;
-}
