+++ src/main/java/games/strategy/triplea/delegate/BidPurchaseDelegate.java
@@ -130,12 +130,3 @@ public class BidPurchaseDelegate extends PurchaseDelegate {
     m_hasBid = s.m_hasBid;
   }
 }
-
-
-class BidPurchaseExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = 6896164200767186673L;
-  Serializable superState;
-  int m_bid;
-  int m_spent;
-  boolean m_hasBid;
-}
