+++ src/main/java/games/strategy/triplea/delegate/EndRoundDelegate.java
@@ -363,12 +363,3 @@ public class EndRoundDelegate extends BaseTripleADelegate {
     return null;
   }
 }
-
-
-class EndRoundExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = 8770361633528374127L;
-  Serializable superState;
-  // add other variables here:
-  public boolean m_gameOver = false;
-  public Collection<PlayerID> m_winners = new ArrayList<>();
-}
