+++ src/main/java/games/strategy/triplea/delegate/InitializationExtendedDelegateState.java
@@ -1,0 +1,9 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class InitializationExtendedDelegateState implements Serializable {
+  private static final long serialVersionUID = -9000446777655823735L;
+  Serializable superState;
+  public boolean m_needToInitialize;
+}
