+++ src/main/java/games/strategy/triplea/delegate/MoveDelegate.java
@@ -625,14 +625,3 @@ public class MoveDelegate extends AbstractMoveDelegate {
     m_PUsLost.add(t, amt);
   }
 }
-
-
-class MoveExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = 5352248885420819215L;
-  Serializable superState;
-  // add other variables here:
-  public boolean m_firstRun = true;
-  public boolean m_needToInitialize;
-  public boolean m_needToDoRockets;
-  public IntegerMap<Territory> m_PUsLost;
-}
