+++ src/main/java/games/strategy/triplea/delegate/PoliticsDelegate.java
@@ -600,11 +600,3 @@ public class PoliticsDelegate extends BaseTripleADelegate implements IPoliticsDe
     }
   }
 }
-
-
-class PoliticsExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = -3829812751864156598L;
-  Serializable superState;
-  // add other variables here:
-  // public HashMap<ICondition, Boolean> m_testedConditions = null;
-}
