+++ src/main/java/games/strategy/triplea/delegate/PoliticsExtendedDelegateState.java
@@ -1,0 +1,10 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class PoliticsExtendedDelegateState implements Serializable {
+  private static final long serialVersionUID = -3829812751864156598L;
+  Serializable superState;
+  // add other variables here:
+  // public HashMap<ICondition, Boolean> m_testedConditions = null;
+}
