+++ src/main/java/games/strategy/triplea/delegate/PurchaseExtendedDelegateState.java
@@ -1,0 +1,10 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class PurchaseExtendedDelegateState implements Serializable {
+  private static final long serialVersionUID = 2326864364534284490L;
+  Serializable superState;
+  // add other variables here:
+  public boolean m_needToInitialize;
+}
