+++ src/main/java/games/strategy/triplea/delegate/RandomStartDelegate.java
@@ -264,11 +264,3 @@ public class RandomStartDelegate extends BaseTripleADelegate {
     }
   }
 }
-
-
-class RandomStartExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = 607794506772555083L;
-  Serializable superState;
-  // add other variables here:
-  public PlayerID m_currentPickingPlayer;
-}
