+++ src/main/java/games/strategy/triplea/delegate/SpecialMoveExtendedDelegateState.java
@@ -1,0 +1,9 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class SpecialMoveExtendedDelegateState implements Serializable {
+  private static final long serialVersionUID = 7781410008392307104L;
+  Serializable superState;
+  public boolean m_needToInitialize;
+}
