+++ src/main/java/games/strategy/triplea/delegate/TechActivationDelegate.java
@@ -158,10 +158,3 @@ public class TechActivationDelegate extends BaseTripleADelegate {
     return null;
   }
 }
-
-
-class TechActivationExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = 1742776261442260882L;
-  Serializable superState;
-  public boolean m_needToInitialize;
-}
