+++ src/main/java/games/strategy/triplea/delegate/TechActivationExtendedDelegateState.java
@@ -1,0 +1,9 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class TechActivationExtendedDelegateState implements Serializable {
+  private static final long serialVersionUID = 1742776261442260882L;
+  Serializable superState;
+  public boolean m_needToInitialize;
+}
