+++ src/main/java/games/strategy/triplea/delegate/TechnologyDelegate.java
@@ -431,12 +431,3 @@ public class TechnologyDelegate extends BaseTripleADelegate implements ITechDele
     return ITechDelegate.class;
   }
 }
-
-
-class TechnologyExtendedDelegateState implements Serializable {
-  private static final long serialVersionUID = -1375328472343199099L;
-  Serializable superState;
-  // add other variables here:
-  public boolean m_needToInitialize;
-  public HashMap<PlayerID, Collection<TechAdvance>> m_techs;
-}
