+++ src/main/java/games/strategy/triplea/delegate/UserActionExtendedDelegateState.java
@@ -1,0 +1,8 @@
+package games.strategy.triplea.delegate;
+
+import java.io.Serializable;
+
+class UserActionExtendedDelegateState implements Serializable {
+  private static final long serialVersionUID = -7521031770074984272L;
+  Serializable superState;
+}
